Komikcast (ID): Refactor DTOs and clean up code
- Rename KomikCastDto.kt to Dto.kt.
- Convert data classes to regular classes in DTOs.
- Move mapping and page list generation logic into DTO classes.
- Remove redundant @SerialName annotations.
- Centralize URL slug/index extraction in Dto.kt.
- Clean up KomikCast.kt by removing redundant logic and using DTO methods.
- Apply spotless formatting.
