+++ src/id/komikcast/src/eu/kanade/tachiyomi/extension/id/komikcast/KomikCast.kt
@@ -79,16 +79,16 @@ class KomikCast : HttpSource() {
 
     override fun searchMangaParse(response: Response): MangasPage = parseSeriesListResponse(response)
 
-    override fun getMangaUrl(manga: SManga): String = "$baseUrl/series/${manga.getSlug()}"
+    override fun getMangaUrl(manga: SManga): String = "$baseUrl/series/${manga.getSlug(baseUrl)}"
 
-    override fun mangaDetailsRequest(manga: SManga): Request = GET("$apiUrl/series/${manga.getSlug()}", headers)
+    override fun mangaDetailsRequest(manga: SManga): Request = GET("$apiUrl/series/${manga.getSlug(baseUrl)}", headers)
 
     override fun mangaDetailsParse(response: Response): SManga {
         val result = response.parseAs<SeriesDetailResponse>()
         return result.data.toSManga()
     }
 
-    override fun chapterListRequest(manga: SManga): Request = GET("$apiUrl/series/${manga.getSlug()}/chapters", headers)
+    override fun chapterListRequest(manga: SManga): Request = GET("$apiUrl/series/${manga.getSlug(baseUrl)}/chapters", headers)
 
     override fun chapterListParse(response: Response): List<SChapter> {
         val result = response.parseAs<ChapterListResponse>()
@@ -97,21 +97,18 @@ class KomikCast : HttpSource() {
     }
 
     override fun getChapterUrl(chapter: SChapter): String {
-        val (slug, chapterIndex) = chapter.getSlugAndIndex()
-        return "$baseUrl/series/$slug/chapters/$chapterIndex"
+        val (slug, chapterIndex) = chapter.getSlugAndIndex(baseUrl)
+        return "$baseUrl/series/$slug/chapter/$chapterIndex"
     }
 
     override fun pageListRequest(chapter: SChapter): Request {
-        val (slug, chapterIndex) = chapter.getSlugAndIndex()
+        val (slug, chapterIndex) = chapter.getSlugAndIndex(baseUrl)
         return GET("$apiUrl/series/$slug/chapters/$chapterIndex", headers)
     }
 
     override fun pageListParse(response: Response): List<Page> {
         val result = response.parseAs<ChapterDetailResponse>()
-        val images = result.data.data.images ?: emptyList()
-        return images.mapIndexed { index, imageUrl ->
-            Page(index, "", imageUrl)
-        }
+        return result.data.toPageList()
     }
 
     private fun parseSeriesListResponse(response: Response): MangasPage {
@@ -139,16 +136,5 @@ class KomikCast : HttpSource() {
             .build()
 
         return GET(page.imageUrl!!, newHeaders)
-    }
-
-    private fun SManga.getSlug(): String = "$baseUrl$url".toHttpUrl().pathSegments[1]
-
-    private fun SChapter.getSlugAndIndex(): Pair<String, String> = if (url.startsWith("/chapter/")) {
-        val slug = url.substringAfter("/chapter/").substringBefore("-chapter-")
-        val chapterIndex = url.substringAfter("-chapter-").substringBefore("-bahasa-")
-        slug to chapterIndex
-    } else {
-        val path = "$baseUrl$url".toHttpUrl().pathSegments
-        path[1] to path[3]
     }
 }
